Refactor regex
